backend/requirements.txt
Django>=5.0
psycopg[binary]>=3.2
python-dotenv
djangorestframework>=3.15
django-cors-headers
djangorestframework-simplejwt